requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
markdown>=3.4.0
python-magic>=0.4.27
beautifulsoup4>=4.9.3 